Moved Artist.kt to entities package, and modified imports
