+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/data/external/artist/artists/ArtistExternalServiceImpl.kt
@@ -2,7 +2,7 @@ package ayds.lisboa.songinfo.moredetails.fulllogic.data.external.artist.artists
 
 import ayds.lisboa.songinfo.moredetails.fulllogic.data.external.artist.LastFMAPI
 import ayds.lisboa.songinfo.moredetails.fulllogic.data.external.artist.LastFMToArtistResolver
-import ayds.lisboa.songinfo.moredetails.fulllogic.domain.Artist
+import ayds.lisboa.songinfo.moredetails.fulllogic.domain.entities.Artist
 
 
 internal class ArtistExternalServiceImpl (
