+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/data/local/sqldb/ArtistLocalStorage.kt
@@ -1,6 +1,6 @@
 package ayds.lisboa.songinfo.moredetails.fulllogic.data.local.sqldb
 
-import ayds.lisboa.songinfo.moredetails.fulllogic.domain.Artist
+import ayds.lisboa.songinfo.moredetails.fulllogic.domain.entities.Artist
 
 interface ArtistLocalStorage {
     fun saveArtist(artist: String?, info: String?,url: String)
