+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/data/local/sqldb/CursorToArtistLocal.kt
@@ -1,7 +1,7 @@
 package ayds.lisboa.songinfo.moredetails.fulllogic.data.local.sqldb
 
 import android.database.Cursor
-import ayds.lisboa.songinfo.moredetails.fulllogic.domain.Artist
+import ayds.lisboa.songinfo.moredetails.fulllogic.domain.entities.Artist
 
 interface CursorToArtistLocal {
